+++ slae-project/slae-project/Graphic/GraphicalLogic.cs
@@ -133,106 +133,5 @@ namespace slae_project
             cursorP.y = initP.y;
         }
     }
-    /// <summary>
-    /// The main form class.
-    /// </summary>
-    public partial class SharpGLForm : Form
-    {
-        GraphicData GD = new GraphicData();
-        
-        /// <summary>
-        /// Initializes a new instance of the <see cref="SharpGLForm"/> class.
-        /// </summary>
-        public SharpGLForm(bool Type)
-        {
-            InitializeComponent();
 
-            GD.ItisATest();
-        }
-
-        /// <summary>
-        /// Handles the OpenGLDraw event of the openGLControl control.
-        /// </summary>
-        /// <param name="sender">The source of the event.</param>
-        /// <param name="e">The <see cref="RenderEventArgs"/> instance containing the event data.</param>
-        private void openGLControl_OpenGLDraw(object sender, RenderEventArgs e)
-        {
-            //  Get the OpenGL object.
-            OpenGL gl = openGLControl.OpenGL;
-
-            //  Clear the color and depth buffer.
-            gl.Clear(OpenGL.GL_COLOR_BUFFER_BIT | OpenGL.GL_DEPTH_BUFFER_BIT);
-
-            //  Load the identity matrix.
-            gl.LoadIdentity();
-
-            GD.RealDraw(openGLControl);
-        }
-
-        /// <summary>
-        /// Handles the OpenGLInitialized event of the openGLControl control.
-        /// </summary>
-        /// <param name="sender">The source of the event.</param>
-        /// <param name="e">The <see cref="System.EventArgs"/> instance containing the event data.</param>
-        private void openGLControl_OpenGLInitialized(object sender, EventArgs e)
-        {
-            //  TODO: Initialise OpenGL here.
-
-            //  Get the OpenGL object.
-            OpenGL gl = openGLControl.OpenGL;
-
-            //  Set the clear color.
-            gl.ClearColor(1.0f, 1.0f, 1.0f, 1.0f);
-        }
-
-        /// <summary>
-        /// Handles the Resized event of the openGLControl control.
-        /// </summary>
-        /// <param name="sender">The source of the event.</param>
-        /// <param name="e">The <see cref="System.EventArgs"/> instance containing the event data.</param>
-        private void openGLControl_Resized(object sender, EventArgs e)
-        {
-            ClearBuffer();
-            //  TODO: Set the projection matrix here.
-
-            //  Get the OpenGL object.
-            OpenGL gl = openGLControl.OpenGL;
-            //
-
-            //  Set the projection matrix.
-            gl.MatrixMode(OpenGL.GL_PROJECTION);
-
-            //  Load the identity.
-            gl.LoadIdentity();
-
-            //  Create a perspective transformation.
-            gl.Perspective(60.0f, (double)Width / (double)Height, 0.01, 100.0);
-
-            //  Use the 'look at' helper function to position and aim the camera.
-            gl.LookAt(-5, 5, -5, 0, 0, 0, 0, 1, 0);
-
-            //  Set the modelview matrix.
-            gl.MatrixMode(OpenGL.GL_MODELVIEW);
-
-            ClearBuffer();
-
-            //Решения проблемы не поспевания отрисовки при Resized, но мб можно что еще?
-            openGLControl.Refresh();
-        }
-
-        /// <summary>
-        /// Изобретаем мегочистящую функцию буфера, но SwapBuffers не нашёл.
-        /// </summary>
-        public void ClearBuffer()
-        {
-            OpenGL gl = openGLControl.OpenGL;
-            gl.Clear(OpenGL.GL_COLOR_BUFFER_BIT | OpenGL.GL_DEPTH_BUFFER_BIT);
-            gl.ClearColor(1.0f, 1.0f, 1.0f, 1.0f);
-            gl.LoadIdentity();
-            gl.Clear(OpenGL.GL_COLOR_BUFFER_BIT | OpenGL.GL_DEPTH_BUFFER_BIT);
-            gl.ClearColor(1.0f, 1.0f, 1.0f, 1.0f);
-            gl.LoadIdentity();
-        }
-
-    }
 }
