+++ slae-project/slae-project/Graphic/SharpGLForm.resx
