+++ slae-project/slae-project/slae-project.csproj
@@ -60,15 +60,16 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
+    <Compile Include="Graphic\GraphicalLogic.cs" />
     <Compile Include="Matrix\CoordinateMatrix.cs" />
     <Compile Include="Matrix\IMatrix.cs" />
     <Compile Include="Matrix\Vector.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SharpGLForm.cs">
+    <Compile Include="Graphic\SharpGLForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="SharpGLForm.designer.cs">
+    <Compile Include="Graphic\SharpGLForm.designer.cs">
       <DependentUpon>SharpGLForm.cs</DependentUpon>
     </Compile>
     <EmbeddedResource Include="Form1.resx">
@@ -83,7 +84,7 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="SharpGLForm.resx">
+    <EmbeddedResource Include="Graphic\SharpGLForm.resx">
       <DependentUpon>SharpGLForm.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="Properties\Settings.settings">
